config.tf/Documentation: tectonic_etcd_servers expects hostnames/IPs (#115)
The only platform that uses `tectonic_etcd_servers` is AWS, more
specifically the `platforms/aws/etcd` module, which either returns
the list of FQDNs for the provisioned etcd nodes, or that list.
For the output to be consistent, we must return the hostnames/IPs
only. Users of that output expects to format that list to add the
port / scheme as necessary, which is more flexible.