Makefile: declare inferred targets as explicit targets (#508)
* Makefile: declare inferred targets as explicit targets

Currently if you invoke `make docs` or `make examples` without deleting
the corresponding files or touching the explicit dependencies, make will
prematurely exit with "Nothing to be done for ...". This makes it tricky
especially for `make examples`, as there are no explicit prerequisites.

This fixes it by specifying the docs/examples targets and thus marking them as
explicit. Additionally this list of targets is marked phony such that
they are always executed.

* Documentation/examples: regenerate