Jenkinsfile: add BM smoke tests and run AWS/BM smoke tests in parallel
* Stop using the Docker agent. It can only be specified at the top-level
or per stage, however, it is impossible to run stages in a parallel()
block. We therefore need to use a local solution instead:
withDockerContainer. Furthermore, the Docker agent does not scale across
multiple nodes, which we need to do for AWS / BM testing. Additionally,
the BM tests can’t run in a Docker container as it is itself launching
other containers and interacting with systemd/libvirt/networking.
* Add a retry loop around AWS’s destroy step in the failure hook.
* Add a variable to store the Tectonic builder image name/version.
* Add a variable for the Quay credentials for clarity.