tests/smoke/aws: enable destroy retry for vpc test
This commit enables the cleanup destroy block to run for private VPC
test clusters even if `create-vpc` was not run in the same shell.

To test:
```
docker run --rm -it -v $PWD:/go/src/tectonic-installer \
    -e AWS_ACCESS_KEY_ID -e AWS_SECRET_ACCESS_KEY \
    --device=/dev/net/tun --cap-add=NET_ADMIN \
    -w /go/src/tectonic-installer quay.io/coreos/tectonic-builder:v1.15 /bin/bash
./tests/smoke/aws/smoke.sh destroy vars/aws-vpc.tfvars
```
Verify that Terraform does not hang while waiting for input. Instead,
Terraform should error because it could not find the tfstate file in
question, which is expected unless the cluster was created. The benefit
is that now, as long as the cluster was created, it does not matter if
`create-vpc` was run or not.