modules/tectonic,tests/smoke: make stats URL configurable
This commit makes the stats reporting URL configurable so that all test
clusters will report statistics to the staging stats database. Although
we already filter out all test data from the production database, it is
useful to submit statistics to the staging database so that we can
happily automate tests that fetch data from that database without
interfering with or causing problems in prod.