installer ui: Separate Tectonic status into separate API call
Following on the last commit, this performs the same separation of
code into two API calls on the frontend side.

A separate function for the call is made in server.js with new
handlers to call it on an interval the same way, and the reducer
is changed to merge statuses in order to allow separate success
dispatches to merge together. This ends with a state which is
identical to that of the single API call, meaning no change to the
React components is needed.

Also, a small section had to be added to the backend function to
cause it to return a 404 before the cluster has been committed,
reflecting the behavior of the Terraform status endpoint, as this
is used by the frontend for routing purposes.