installer ui: Fail Tectonic if Terraform fails
A minor change from the previous commit. Simply adds the same
failure condition to the "Starting Tectonic Console" step as is
already on the Terraform step; thus, if Terraform fails, it causes
Tectonic to show as failed as well, preventing the potentially
confusing state where Terraform fails but Tectonic may appear
to have succeeded.