installer: Add remaining services to status and UI
This commit adds the bulk of the logic to the status function
regarding getting the statuses of the components, as well as to
the UI in displaying these statuses.

It does not reflect the final UI design. Also, the status endpoints
will be merged back into one endpoint in a future commit.