installer api: prevent tectonic status on delete
I noticed previously that when terraform is deleting the cluster,
at a certain point it will cause the Tectonic status handler to
hang, most likely due to timeouts on contacting the Kubernetes API
or a similar problem. This change ensures that tectonic status is
not run while Terraform is deleting the cluster, to prevent this
hanging.