tests: Add upstream terraform trigger
We are currently using a custom terraform version for our default CI
pipeline (See https://github.com/coreos/tectonic-installer/pull/1247).
As end users of Tectonic installer use upstream terraform we need to
test with upstream terraform as well. We are building and publishing the
Tectonic builder image both with CoreOS Terraform as well as upstream
Terraform. This is done via docker `--build-arg` `TERRAFORM_URL`. CoreOS
Terraform is used for normal PR and branch tests, upstream Terraform is
used once per day on master.

* Add `--build-arg` `TERRAFORM_URL` to Tectonic builder Docker file
* Update documentation in tectonic-builder/README.md
* Add upstream-terraform-trigger.groovy pipeline job

With this commit v1.36 and v1.36-upstream-terraform of the Tectonic
builder images are pushed to quay.io/coreos/tectonic-builder. There
are no functional changes, but changes to the build order in the
Dockerfile.

You can find the Jenkins job
[here](https://jenkins-tectonic-installer.prod.coreos.systems/job/upstream-terraform-trigger/).