frontend: Add progress bars to Start Installation page
Adds progress bars for both the AWS Terraform apply step and for the
Starting Tectonic step for both AWS and bare metal.

The Terraform progress is estimated by parsing the log output. This PR
uses a very simple approach with the intention of improving the progress
bar accuracy in a future PR.

The Tectonic progress is based on the sub-steps, except that the
progress bar will not move backwards of a sub-step flips off again.