tests: fix GUI test
The GUI tests are failing because the are using one of the output files
from a different frontend test rather than their own configuration. This
GUI test modifies the configuration file to suite its needs, which will
always be a game of catchup. This fixes the issue for the time being but
it will need more refactoring.