terraform: avoid resource churn for EBS volumes
Possibly related to this issue with the Terraform AWS provider:
https://github.com/terraform-providers/terraform-provider-aws/issues/1521

Terraform or the AWS API report `0` IOPS for `gp2` volumes after creation.
If I create a `gp2` volume with these modules, every subsequent run of
Terraform will attempt to destroy and recreate the LC in order to make
IOPs match. This change sets IOPS to 0 for non-`io1` type volumes.

Since the number of IOPS for `gp2` volume is based on the size, Terraform
arguably should only accept the `iops` parameter for `io1` volumes.

Interestingly, for the etcd nodes, this problem does not occur because
the Terraform state shows 100 IOPS even for `gp2` volumes.