tests: Add hint to README for license and pull secret path
To improve usability of runnning the smoke tests locally, we run them
inside a docker container. The Tectonic installer repository will be
mounted into the container before executing the tests. In order for the
installer to read the *pull-secret* and *license* file, they have to be
in the repository folder, otherwise they can not be read by the
installer during runtime.

This commits adds a hint to the README to point this fact out.