modules/ignition: introduce runtime-mappings.yaml
This introduces a runtime-mappings.yaml to be used by tectonic-torcx.

In particular, this file is consumed by k8s-node-bootstrap.service
as a fallback to determine the docker version to install when the
api-server is unreachable (e.g. while installing a brand new cluster).
If an api-server is available, an up-to-date version of this data
will be sourced from a config-map instead as a primary source.