modules/ignition: ensure --cloud-provider is set
The default behavior of the kubelet's `--cloud-provider` flag is
changing in the upcomint k8s releases:

* v1.8 - auto detection of cloud provider using cAdvisor is deprecated but
still remains the default.
* v1.9 - auto detection of cloudprovider is still supported but no longer
the default, the new default will be no provider.
* v1.10 - auto detecting cloud provider is removed

For consistent non-autodetection behavior we need to ensure that we
always set the cloud-provider flag, even if it is explicitly set to no
provider.