tests: Improve wait_for_bootstrapping
The Tectonic installer CI had repeatedly failed with timeouts waiting
for the bootkube and tectonic service to finish.

This patch includes the following improvements:
- The bootkube and tectonic systemd unit is only deployed on one master
node. Instead of relying on the fact, that we choose the right master
node to check if the services finished, we check all of them and return
true if the services succeeded on one of them
- Instead of checking for success via:
`systemctl is-active <service>`, we check whether the
`init_<service>.done` file exists. `is-active` returns non zero when the
bootstrap succeeded but the node restarted in the meantime. The
`init_<service>.done` file is only written when the service exits with a
non-zero exit code [1]

[1] See `ExecStartPost`
https://www.freedesktop.org/software/systemd/man/systemd.service.html