tests: Report specific smoke test status to GitHub (#2082)
* tests: Report specific smoke test status to GitHub

Instead of reporting overall status to Github, this patch introduces
more detailed status reports:

1. `basic-tests` includes all basic tests in the first stage of our
pipeline. This status will be required to be green. In case it is red, a
merge of the PR is disabled.
2. `gui-tests` represents the result of the GUI tests stage. It is not
mendatory for merge. It will only run when the `run-gui-tests` label is
set.
3. `smoke-tests/<test-name>` is going to be reported per smoke test.
None of these status reports are mandatory for merge. These tests are
only run when the `run-smoke-tests` label is set.
4. `continuous-integration/jenkins/pr-merge` is the overall status of
the entire pipeline. This is not mandatory. I am looking into
deactivating this in total, but haven't found a solution yet.

* tests: Report bare metal status to GitHub

* tests: Make sure try catch block surrounds entire test execution

* tests: Restrict scope of credentials injection

* tests: Move smoke unstash step into docker container

* tests: Remove stashing logic

According to the Jenkins documentation `stash` should only be used for
small amounts of data [1]. To prevent `java.io.IOException: Failed to
extract repository.tar.gz` this patch reduces the stash scope to the
npm_modules, the installer binary and the smoke test binary. The
remaining repository files are loaded via standard git checkouts.

[1]
https://jenkins.io/doc/pipeline/steps/workflow-basic-steps/#code-stash-code-stash-some-files-to-be-used-later-in-the-build