frontend: Delay /aws/iam-roles API call until Credentials page completed
We don't need a region to call /aws/iam-roles, so the call we being made
as soon as the AWS id and secret key were provided. Unfortunately, that
could result in lots of unnecessary API calls as the values are entered.
To avoid this, wait until the whole form (include the region dropdown)
is completed.