frontend: Remove unnecessary entries when saving progress file
We shouldn't be relying on any of these entries when restoring progress,
and removing them ensures that we are not.

error, error_async, extraError and ignore don't need to be saved because
validation is run on all fields when the progress file is loaded, which
recalculates all of these values.

inFly and extraInFly indicate whether an API is in progress, which
doesn't make sense to store in the progress file.