modules/aws: detect-master assumes lb
This commit fixes issue #2419.

This issue occurs because we recently added -o pipefail to the
detect-master.sh script. This causes a previously undetected issue to
surface and break bootstrapping for clusters with no private-endpoints.

API_HEALTHY=$(aws elb describe-instance-health --region="$REGION"
--load-balancer-name "$CLUSTER_NAME-int" | jq -r '[ .InstanceStates[] |
select(.State | contains("InService")) ] | length > 1')
is hardcoded to use the internal loadbalancer. Without pipefail, this
aws command would 404 and the error would get swallowed by the pipe.
API_HEALTHY would never be true, even if the API was actually healthy,
so the rest of the script would run.

Now, since the pipefail option is set, that command correctly errors out
and the script execution is aborted.

We simply need to template the loadbalancer name based on whether
private endpoints are enabled or disabled.