frontend: Fix display of error message when AWS credentials are invalid
This error was already rendered as a field validation error on the
regions dropdown, but that meant it was only shown if the dropdown was
marked as "dirty" in Redux, which doesn't happen until its value is
first set.

Explicitly rendering the AWS API error solves this problem, but it does
mean that the error message can now possibly be displayed twice. Solving
that will require a larger code change, so this is an initial quick fix.