tests: lock azure provider version
The Azure Terraform provider version is locked for the Azure platform,
however, it is not specified for the test fixtures that create some
Azure resources before the actual Tectonic install begins. This means
that we cannot reliably download the needed version of the Azure
provider during the fixture part of the Azure smoke tests. This is
causing issues since version 1.0.0 of the Azure provider was just
released but an incorrect checksum was published, causing downloads to
fail.