tests/frontend: Add a page refresh on the last wizard page
This test confirms that field values are preserved during a page
refresh. If a value is not being preserved, the final TF vars diff step
will fail as long as that value is not the same as the field's default
value.

Also changes 2 test values that were the same as the field's defaults.