modules/ignition,vmware: Enable install behind HTTP proxy
This addition will allow a cluster to be installed behind an HTTP proxy
on VMware.  The changes have been implemented in the ignition module to
allow it to be readily added to the other platforms.

Note: this only addresses the env vars needed on the CL nodes.  There
are tectonic components - like tectonic-channel-operator - that also
need these env vars to properly function.  That will be addressed
separately.