bazel: create tarball target (#2750)
This commit adds tarball targets for Bazel.

We can now create tarballs using:

```sh
bazel build tarball
```

This will create a gzip compressed tarball named tectonic.tar.gz. The
contents of the tarball will be nested in a directory named `tectonic`.

To create a tarball with a version string in the directory name, we can
build the tarball using:

```sh
export VERSION=<some version>
bazel build tarball --action_env=VERSION
```

In this case, the file structure will be nested in a directory called
`tectonic_<some version>`.

This commit also makes the installer binaries pure go so that they are
statically linked, which is nicer for containers.