frontend: Rename REMOVE_AT action to REMOVE_FIELD_LIST_ROW and simplify
This Redux action is now specifically for removing a row from a field
list, so rename it to make that clear.

Simplify the reducers by assuming they take a field list id and an
index, not an arbitrary path. Also stop clearing any associated error
data, since a REMOVE_FIELD_LIST_ROW action will need to trigger a
re-validation of the field list anyway.