frontend: Output custom subnet tfvars if Advanced Networking is opened
It used to be that AWS_CONTROLLER_SUBNETS and AWS_WORKER_SUBNETS were
marked as dirty as soon as the Advanced Networking section was opened,
but that stopped being the case due to AWS VPC and forms changes. We
should therefore track the dirtiness of AWS_ADVANCED_NETWORKING itself
and use that to decide whether to output master and worker custom
subnet tfvars.