platforms/metal: unbundle custom certs
This commit un-bundles the custom CA certificates that are passed to the
Tectonic Installer. This is needed for OpenSSL certificate hashing to
work. Without this, CAs cannot be anchored. This commit adds an
intermediate templating step to the Matchbox profile generation where
custom CA certs are inserted. Because of the fact that the Tectonic
Installer modifies the `update-ca-certificates.service` unit via a
drop-in, certificates are rehashed on every reboot so the custom CA
certs are rehashed automatically and are avaiable on the first boot to
programs compiled against OpenSSL.