frontend: Add additional info for AWS VPCs and subnets
For VPCs, add CIDR to dropdown labels and add CIDR tooltip with number
of IP addresses and address range.

For subnets, add Name (from AWS tag) to dropdown labels and add CIDR
tooltip with number IP addresses, address range and number of available
addresses.

Change label format of both dropdowns to "<CIDR> (<NAME> | <ID>)", which
matches the format in AWS Console.

Change Select to accept an "options" array instead of an
"availableValues" object.

Change Deselect to also output the field label, which makes things
easier because the deselect checkbox is expected to be rendered
alongside the label and it's easier to ensure they use matching id /
htmlFor attributes if they are rendered together.

Rename fieldName to deselectId and change CIDRRow to take deselectId
instead of a "selectable" boolean.