frontend: Fix AWS regions duplicated error and add loading indicator
Move the AWS region dropdown / region load error message to a separate
component.

Don't display the AWS regions dropdown if there was a load error. This
fixes a situation where the error was displayed both as a regions field
error and as a general form error.

Display a loading indicator instead of the regions dropdown while the
request is pending.