modules/container_linux: avoid shell
Due to the way Terraform ternary expressions are evaluated, even if you
specify a Container Linux version to install, Terraform will invoke the
container_linux module external data source and try to run `curl` and
make an HTTP request. If the host is offline or does not have `curl`
installed, then this module will fail. This commit enables this module
to be circumvented by explicitly defining the Container Linux version to
install. Furthermore, this enables non-unix hosts to use the module.