tests: Optionally trigger pipeline with specific git ref
In order for the release automation pipeline to consume the results of
the Tectonic installer pipeline, we have to ensure, that both are
configured with the same Tectonic installer git ref (commit).

This patch adds a Jenkins pipeline parameter to the Tectonic installer
pipeline to trigger a build with a specific git ref.