tests: Adjust log-analyzer-copy to missing build folder
With the new Tectonic cli flow, the cluster build folder does not
exist anymore. Instead the cli creates a folder on the top level with
the name of the cluster, containing all relevant files.

For now we need to support both flows. For AWS the Tectonic CLI is used.
For Azure and all other platforms we still use plain terraform.