frontend: Fix AWS custom subnet validation bug when AWS region changes
Change the action so that it always overwrites the AWS subnets so that
the subnet keys (which include the AWS AZs) are updated when the AWS
region is changed.

Unfortunately, this means that any user entered custom subnet CIDRs will
be lost if the page is refreshed, so the Nightwatch tests are also
updated to reflect that. Fixing this will require a larger code change.