config.tf,steps: cleanup tectonic_ca
The tectonic_ca_cert option is not just for setting the Tectonic Console
server certificate but rather is used as a root certificate from which
to generate all cluster certificates. This commit cleans up the options
a little bit. More cleanup will follow once all TLS asset generation is
moved to the CLI.