scripts/maintenance/tag-route53-hosted-zones: Drop '-e' from echo
The 'echo -e' calls landed with the script in 82bdd9fe
(installer/scripts: AWS tag and delete scripts, 2017-06-28,
coreos/tectonic-installer#1239).  But $tags doesn't actually need any
backslash expansion:

  $ date_string='2000-01-01'
  $ tags="[{\"Key\":\"expirationDate\",\"Value\":\"$date_string\"}]"
  $ echo "${tags}"
  [{"Key":"expirationDate","Value":"2000-01-01"}]

Perhaps the original concern was over the \" in the tags definition,
but those are expanded by the shell during that initialization.  By
the time we get around to invoking echo, there are no backslashes left
(unless the user has injected some via --date-override, and we don't
want to support that ;).