README: Document console URL
I spent way to long trying to figure out why
https://${CLUSTER_NAME}.${BASE_DOMAIN} was giving me an "Application
is not available" error.

While I'm adding this to the AWS notes in the README, also touch the
libvirt notes to add braces, so it's easier to see that '-api' is a
literal, and not part of the variable name.

Chance Zibolski suggested that the -api hostname suffix and/or 6443
port are because production deploys might want to put an elastic load
balancer [1] in front of the current API host for production deploys.

[1]: https://aws.amazon.com/elasticloadbalancing/