Documentation/variables: Drop stale docs
These used to be generated with terraform-docs, but the associated
Makefile rule was removed without replacement in d61abd48 (*: cleanup
bazel rules, 2018-03-26, coreos/tectonic-installer#3137).
terraform-examples seems to have been similar, and its last Makefile
reference was also removed in d61abd48.  We may want to return to
autogenerating examples/*.yaml at some point, but for now, stop
claiming that we are autogenerating these.