steps: introduce bootstrap node
This also removes the rm-assets service since the bootstrap node itself
will be removed after the cluster is bootstrapped. Because rm-assets was
the only service that was actually enabled, tectonic.service instead
needs to be disabled.

config.tf is not bumped here because there is no corresponding change in
TNC. Even if the master config was served from TNC, there would be a
race condition between the etcd cluster becoming healthy and the master
requesting its config. Rather than attempting to address this race, I'm
ignoring it since the etcd members are going to be moved to the master
soon, which will remove the race.