pkg/installerassets: Use the generic asset package
... to create installer assets.  This asset graph is currently pretty
sparse while we work out the proof-of-concept.

The 'graph' command is a bit different now that parents are dynamic.
Instead of printing the full possible graph, it prints the actual
graph for a particular asset state.  If you render an asset graph with
'openshift-install --dir=whatever create assets', edit a file, and run
'openshift-install --dir=whatever graph', you'll get a graph for your
particular asset store.

I'm also auto-coloring nodes based on their directory name, using
hue-saturation-value colors to get a rainbow of pale backgrounds.

I'm reaching in and adjusting the saturation for
tls/kubelet-client.crt because it is special with its thirty-minute
validity.  While the other assets, especially the other TLS assets,
should be safe to use for multiple cluster invocations, users will
almost certainly want to remove and regenerate the kubelet client cert
for each new cluster (or drop in their own client cert with a longer
validity).