Create a pty by default
RHCOS defaults to use ttyS0 (the first serial port) as the console for init
(and things after init). Since our domains do not have a pty defined these
go into the void. With this change one can use `virsh console` and see the
messages.  (If RHCOS can figure out how to send output to both tty0 and ttyS0
we wouldn't need this)