steps: Assign master and etcd security groups to both node classes
This reduces isolation between the master and etcd nodes, and grants
the etcd nodes more access to workers than they had before.  But we're
going to move etcd onto the master nodes soon, which will have the
same effect.  Hopefully this relaxed isolation helps us avoid:

  $ curl -s https://storage.googleapis.com/origin-ci-test/pr-logs/pull/openshift_installer/150/pull-ci-origin-installer-e2e-aws/546/build-log.txt | grep Ginkgo
    |  Ginkgo timed out waiting for all parallel nodes to report back!  |
  Ginkgo ran 1 suite in 10m6.626061944s

which we've been seeing recently.  Clayton suspects the e2e-aws tests
are trying to get metrics from the kubelets, and hanging on the etcd
kubelet because access from master -> etcd was restricted.  I'm not
clear why we've only been seeing this issue for the last week though.