tests/run: Shift AWS_REGION config down to the AWS role setup
This collects our AWS configuration to make it easier to toggle it
on/off all of our AWS-specific stuff, which we'll need as we add
support for libvirt (and potentially other smoke-test backends) in the
future.

I've also added an `aws configure get region` fallback, so the logic
is:

1. If the user has set AWS_REGION, use what they set.  Otherwise...
2. If we can extract a region from the user's config (with
   `aws configure get region`), use that.  Otherwise...
3. Fall back to eu-west-1.

Step 2 is new in this commit, and makes life a bit more convenient for
callers who have a preferred region (because they have their
preference respected without having to bother setting `AWS_REGION`).

The ${parameter:-[word]} syntax is specified in [1].

[1]: http://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#tag_18_06_02