Remove the 'dns' options from libvirt
This will allow the libvirt nodes to use the hypervisor's DNS instead of
having to hard code some upstream DNS server. Previously this was a
problem if your local machine pointed to the libvirt dnsmasq for DNS
resolution of the nodes. Any unknown address would cause an inifnite
loop. Now, the libvirt dnsmasq will respond that the name is unknown
instead of forwarding it to the hypervisor's DNS server.

Since i can't see a reason for the DNS option, I take it out. Now it
'just works'.