installer/pkg/config-generator/ignition: Ignore etcd node pools
Until we can re-land [1] (which is currently reverted by [2]).  With
the etcd node pool still defined in the release template, we're
getting [3]:

  time="2018-08-27T18:50:53Z" level=fatal msg="unrecognized role: etcd"

With this commit, we ignore that entry (just as Go's JSON and YAML
unmarshallers ignore other properties which aren't reflected in the
target Go structures).  A better long-term fix would probably be
adding more structure to node pools, but I'm punting on that for now.

[1]: https://github.com/openshift/release/pull/1274
[2]: https://github.com/openshift/release/pull/1284
[3]: https://storage.googleapis.com/origin-ci-test/pr-logs/pull/openshift_installer/168/pull-ci-origin-installer-e2e-aws/710/build-log.txt