pkg/asset/installconfig/ssh: Discover ~/.ssh/*.pub
With this commit, we get a new prompt like:

  SSH Public Key:
  Enter an empty string, your public key (e.g. 'ssh-rsa AAAA...'), or one of the following numbers:
  1: /home/trking/.ssh/id_rsa.pub

This commit also adds an "is this SSH pubkey valid?" check based on
our previous validation.

The BUILD.bazel change was generated with:

  $ bazel run //:gazelle

The testify addition there is catching up with 345d8431 (asset: Only
persistent data to disk for targetable assets, 2018-08-30, #196).