steps: consolidate machines and topology
Machines and the rest of topology can be created at the same time now
that CNAMEs are no longer used with the masters. After the new "infra"
step executes, all of the infra required for the running cluster is
created. It will remain in a deadlocked state until a bootstrap node
breaks the dependency loop (which happens in the following step).