*: Replace PullSecretPath with PullSecret
Instead of passing the pull secret around as a path, pass it around as
a JSON string.  This makes it easier to embed in Kubernetes, since
we're punting the file-reading to callers at config-YAML-creation
time.

Store it as a string (like the SSH pubkey) instead of parsing it out
into a more detailed structure, because we expect to consume it as an
opaque string (just pass it to the registry without peaking inside).

I've left some deprecated handling for folks who are still using
pullSecretPath in their YAML.  I'll file a follow-up pull request to
drop it once we get the CI template in openshift/release updated to
use pullSecret.