modules/aws/bootstrap: Pull AWS bootstrap setup into a module
This will make it easier to move into the existing infra step.

The module source syntax used in the README is documented in [1,2,3],
and means "the modules/aws/ami subdirectory of the
github.com/openshift/installer repository cloned over HTTPS", etc.

I don't think I should need the wrapping brackets in:

  vpc_security_group_ids = ["${var.vpc_security_group_ids}"]

but without it I get [4]:

  Error: module.bootstrap.aws_instance.bootstrap: vpc_security_group_ids: should be a list

The explicit brackets match our approach in the master and worker
modules though, so they shouldn't break anything.  It sounds like
Terraform still has a few problems with remembering type information
[5], and that may be what's going on here.

I've simplified the tagging a bit, keeping the extra tags unification
outside the module.  I tried dropping the kubernetes.io/cluster/ tag
completely, but it lead to [6]:

  Sep 12 05:08:08 ip-10-0-6-58 hyperkube[4122]: E0912 05:08:08.075211    4122 tags.go:94] Tag "KubernetesCluster" nor "kubernetes.io/cluster/..." not found; Kubernetes may behave unexpectedly.
  Sep 12 05:08:08 ip-10-0-6-58 hyperkube[4122]: F0912 05:08:08.075258    4122 server.go:262] failed to run Kubelet: could not init cloud provider "aws": AWS cloud failed to find ClusterID

The backing code for that is [7,8,9].  From [9], you can see that only
the tag on the instance matters, so I've dropped
kubernetes.io/cluster/... from volume_tags.  Going forward, we may
move to configuring this directly instead of relying on the tag-based
initialization.

[1]: https://www.terraform.io/docs/configuration/modules.html#source
[2]: https://www.terraform.io/docs/modules/sources.html#github
[3]: https://www.terraform.io/docs/modules/sources.html#modules-in-package-sub-directories
[4]: https://storage.googleapis.com/origin-ci-test/pr-logs/pull/openshift_installer/217/pull-ci-openshift-installer-e2e-aws/47/build-log.txt
[5]: https://github.com/hashicorp/terraform/issues/16916#issuecomment-351848789
[6]: https://github.com/openshift/installer/pull/217/#issuecomment-420514281
[7]: https://github.com/kubernetes/kubernetes/blob/v1.11.3/pkg/cloudprovider/providers/aws/tags.go#L30-L34
[8]: https://github.com/kubernetes/kubernetes/blob/v1.11.3/pkg/cloudprovider/providers/aws/tags.go#L100-L126
[9]: https://github.com/kubernetes/kubernetes/blob/v1.11.3/pkg/cloudprovider/providers/aws/aws.go#L1126-L1132