steps/bootstrap: Merge into the infra step
Booting clusters is faster and teardown is more robust if we can
perform everything in a single Terraform step.  With this commit, we
are down to that single step.

It would be nice to drop the terraform_remote_state hack for the
bootstrap Ignition.  But we'll be doing that when we switch to
cmd/openshift-install and its all-Go asset generation, so I'm rolling
with the existing approach for now.