tests/run: Destroy with --continue-on-error
Taking advantage of 23915f2e (installer/*: Add --continue-on-error to
destroy workflow, 2018-09-16, #252) so we can reap at least most of
our resources even if the cluster doesn't come up enough for the
machine API operator to be able to destroy workers.  With various
stages of cluster health:

1. Cluster never comes up at all.
2. Cluster healthy enough to create workers.
3. Cluster healthy enough to destroy workers.

we're only worried about leakage in the space between 2 and 3.
Hopefully there isn't any space there, but without this commit we're
currently leaking resources from 1 as well.