aws/ELBs: merge tnc with api_internal and cleanup
Remove the TNC ELB with and move all associated resources to the
api_internal ELB. This also allows to cleanup the DNS A record and the
security group for the TNC. It also changes the FQDN for the TNC, which
is now the same as for the API, though it remains exclusive to the
internal zone.

Configure the ELB to listen on the TNC port directly.