assets/ignition: add all kubernetes manifests
This adds the missing Kubernetes manifests to the bootstrap Ignition
config. This also simplifies this process by relying on the asset
contents' names instead of hardcoding them. The tests for the bootstrap
Ignition config have been removed because this sort of test is very
difficult to keep up-to-date. As a result, developers tend to blindly
copy and paste to ensure the tests pass. This functionality will be
covered by integration tests anyway.