asset/cluster: Invoke terraform in a temp dir.
In order to be able to use openshift-install to launch multiple
clusters we need to create a working dir for each invocation.

Here we create a tmp dir that has the terraform.tfvars file and
invoke the cluster within the dir.

The dir will be removed after the invocation is done.