data: Setup a new data package
Based on [1].  Generate at release-time with:

  $ go generate ./data/

I've swapped [1]'s tag from 'dev' to 'release' (with the opposite
sense) because we're not going to commit the autogenerated content to
the repository ('go get' support is not worth version-controlling
machine-generated code).

Because dev builds load the data at runtime (instead of compiling it
in at build time), I've added an environment variable to fill the role
of the previous terraform.BaseLocation.

[1]: https://github.com/shurcooL/vfsgen/blob/33ae1944be3fe078a54c597f107e0867da19c713/README.md#go-generate-usage