images/installer: Rewrite tectonic-installer for openshift-install
The golang-1.10 image has everything we need to build now.  I run the
build in a golang-1.10 container using 'FROM ... AS ...' [1] like we
used to.  But now I no longer install packages with yum.  And I use a
recursive COPY [2] to bring the built/fetched binaries over into the
output container, which is based on scratch.

Once we switch openshift/release over to this Dockerfile, we can drop
images/tectonic-installer.

[1]: https://docs.docker.com/engine/reference/builder/#from
[2]: https://docs.docker.com/engine/reference/builder/#copy