images/installer: Copy binaries separately
My old podman 0.7.3 does not support wildcard copies [1], and the old
.../bin -> /bin copy breaks targets where /bin is a symlink to
/usr/bin.  That doesn't matter for scratch, but if you use this same
Dockerfile for, say, openshift/origin-v4.0:base, the clobber means
that /bin/sh doesn't work (because on that CentOS image, /bin/sh
resolves to /usr/bin/bash via /bin -> usr/bin and /usr/bin/sh -> bash
symlinks).

Also set USER [2] to make it clear that the installer doesn't need to
run as root.  Set both a UID and GID because our scratch image won't
have an /etc/passwd or similar to resolve the 1000 user's primary
group.

[1]: https://docs.docker.com/engine/reference/builder/#copy
[2]: https://docs.docker.com/engine/reference/builder/#user