pkg/asset/installconfig/ssh: Add OPENSHIFT_INSTALL_SSH_PUB_KEY_PATH
To make it easier to pass this in when you don't have a shell for:

  $ OPENSHIFT_INSTALL_SSH_PUB_KEY="$(cat path/to/key)" openshift-install install-config

For example, this is useful in CI where we can launch the cluster
without hitting a shell at all.  And it's also a bit more compact for
folks who are reading from files anyway.

I've also added a check for "do we only have one choice?".  If so, I
just pick that value instead of bothering the user when they don't
have a decision to make.