hack/get-terraform: Use 'go env ...' to support other platforms
Jeff Ligon is on OS X.  With this change, we use Go to give the
correct OS and architecture constants for the Terraform URL.  Go,
because Terraform is a Go project, so Hashicorp is building the
binaries using Go's cross-compilation with the Go naming.

If you don't have Go, I'm just assuming you're on amd64 Linux, because
that's what our CI infrastructure is running.