pkg/asset/installconfig: Add _CI_ONLY_STAY_AWAY_OPENSHIFT_INSTALL_AWS_USER_TAGS
With a JSON string containing the intended values.

This is not how we're going to expose this long-term, so I'm using an
embarrassing name and not documenting the enviroment variable.  But I
want this so we can get back to tagging expirationData in CI, without
waiting for asset state <-> disk (de)serialization.